refactored alot
