+++ src/screens/swap/config.ts
@@ -1,4 +1,14 @@
+import { DexTypeEnum } from 'swap-router-sdk';
+
 export const ROUTING_FEE_ADDRESS = 'tz1XYSt74vwVWgixEXdiS4C5t3UvuExRRTZR';
 export const TEZOS_DEXES_API_URL = 'wss://tezos-dexes-api-mainnet.production.madservice.xyz';
 export const ROUTING_FEE_PERCENT = 0.5;
 export const ROUTING_FEE_RATIO = (100 - ROUTING_FEE_PERCENT) / 100;
+
+export const KNOWN_DEX_TYPES = [
+  DexTypeEnum.QuipuSwap,
+  DexTypeEnum.Plenty,
+  DexTypeEnum.LiquidityBaking,
+  DexTypeEnum.QuipuSwapTokenToTokenDex,
+  DexTypeEnum.Youves
+];
