+++ src/screens/swap/swap-form/swap-form-amount-input.tsx
@@ -39,6 +39,8 @@ export const SwapFormAmountInput: FC<Props> = ({
     }
   };
 
+  console.log('SwapFormAmountInput', name);
+
   return (
     <>
       <SwapFormAssetAmountInput
