+++ src/screens/swap/swap-form/swap-form-dropdown.tsx
@@ -77,6 +77,8 @@ export const SwapFormDropdown = ({
     dropdownBottomSheetController.close();
   };
 
+  console.log('SwapFormDropdown', title);
+
   return (
     <>
       <TouchableOpacity
